move the tests to where it makes more sense
